refactor: move Accordion to ArchivedComponents
